Support for SSLRequest responses in pgsql
In postgres, the response to a SSLRequest is one byte response:
N (if SSL is rejected), S (if SSL is accepted). These responses
need to be differentiated from NoticeResponse(N) and ParameterStatus(S),
that are more than one byte responses. Thus, the parser has to expect
for one byte responses only in the case SSLRequest was seen in the
tcp stream.