Added a -test CLI option.
It does the initialization normally, but exists before processing
any packets. The change also make sure we exit with a non zero
exit code in case of issues with the config.

This is meant to be used by the init scripts to check for common
errors before starting in daemon mode. #9