Proper error handling code in the mysqlParseResponse function
The problem was that the initial version of this function assumed
that the message is already a valid MySQL message. While indeed
some basic checks on it were made at the layer below (separating
the message out of the TCP stream), there are still plenty of things
that can go wrong.

This adds error handling in (hopefully) all possible branches and
a set of unit tests that exploit most of those.

Closes #180.