Declare the flag in init() but change its default value afterwards
This makes the flag work for the coverage-enabled system tests but
also have a nice default value depending on the Beat name. And unlike
other possible solutions, it doesn't leak boilerplate into the Beats.

One caveat is that the system tests binary has a different default
value for this flag, but that doesn't matter in our case.